Update style.css
This works on iPad. Center horizontal in portrait and vertical in landscape - looks good, so think we should use this as a temporary solution. I've moved the other CSS to WIP.css
With this css you can also remove the forecast length on iPad 2 day only limit - should work with al lengths.